+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @marica.io/ui
 
+## 0.1.4
+
+### Patch Changes
+
+- - Remove dark mode
+  - Add child slot to input text
+  - Add Rich Text Editor
+  - Init Sheet
+- Updated dependencies
+  - @marica.io/style@0.1.4
+
 ## 0.1.3
 
 ### Patch Changes